+++ src/hooks/dialogs/usePostDialog.tsx
@@ -12,8 +12,15 @@ export const usePostDialog = <
   props: UsePostDialogPropsType<TMutationDto, TMutationOutputDto, TFormType>,
 ): UseFormDialogReturnType<TFormType> => {
   const queryClient = useQueryClient();
-  const { mutationFn, queryKey, onSuccess, onError, mapOut, ...coreProps } =
-    props;
+  const {
+    mutationFn,
+    queryKey,
+    onSuccess,
+    onError,
+    formToDto,
+    mapOut,
+    ...coreProps
+  } = props;
 
   const dialogFn = useMutation<TMutationOutputDto, Error, TMutationDto>({
     mutationFn,
@@ -22,7 +29,7 @@ export const usePostDialog = <
   return useFormDialog<TFormType, TMutationDto>({
     ...coreProps,
     mode: "entity",
-    mapOut,
+    formToDto: formToDto || mapOut,
     onSubmit: async (payload) => {
       try {
         const result = await dialogFn.mutateAsync(payload);